+++ packages/widgetbook/lib/src/args/double_arg.dart
@@ -25,3 +25,28 @@ class DoubleArg extends Arg<double> with SingleFieldOnly {
     );
   }
 }
+
+class NullableDoubleArg extends Arg<double?> with SingleFieldOnly {
+  const NullableDoubleArg(
+    super.value, {
+    super.name,
+  });
+
+  @override
+  Field<double> get field {
+    return DoubleInputField(
+      name: 'value',
+      initialValue: value ?? 0.0,
+    );
+  }
+
+  @override
+  NullableDoubleArg init({
+    required String name,
+  }) {
+    return NullableDoubleArg(
+      value,
+      name: $name ?? name,
+    );
+  }
+}